+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.5.1] - 2017-5-21
+
+### add
+
+- error-function for multiclass classification
+- add LSTM demo for multiclass classification
+
+### Changed
+
+-
+
+
 ## [0.5.0] - 2017-5-20
 ### Changed